pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "encryptor"
version = "0.1.0"
description = "Pre-commit tool to scramble and encrypt repository content"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pycryptodome>=3.20",
    "pyyaml>=6.0",
]

[project.scripts]
encryptor = "encryptor.cli:main"
